fix: Remove int cast in Average.getValue() to preserve 64-bit precision
The Average aggregation function was incorrectly casting the result
to int before returning as long, which truncated values beyond the
32-bit range. This caused incorrect results when:
- Individual values exceeded Integer.MAX_VALUE
- Sum of values exceeded Integer.MAX_VALUE (even if average didn't)

This fix removes the unnecessary cast, preserving full 64-bit precision
in the average calculation.

Added comprehensive test suite covering:
- Large values exceeding Integer.MAX_VALUE
- Large sums where average is in range
- Negative values
- Edge cases (empty, clear, mixed values)

Fixes: CRITICAL bug identified in btrace-core analysis

+++ btrace-core/src/main/java/org/openjdk/btrace/core/aggregation/Average.java
@@ -53,7 +53,7 @@ class Average implements AggregationValue {
     if (count == 0) {
       return 0; // Avoid division by zero
     }
-    return (int) (sum / count);
+    return sum / count;
   }
 
   @Override
